+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]## [0.5.5.2b5]
 ### Fixed
-- ScenarioRunner.get_parameters can handle case where inputs do not (yet) have a 'Parameter' table.
+- ScenarioRunner.get_parameters can handle case where inputs do not (yet) have a 'Parameter' or 'Parameters' tab
 
 ## [0.5.5.2b4] - 2024-04-10
 ### Added